pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent"
version = "0.1.0"
description = "AnimAI agent package"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.0.0",
    "python-dotenv>=1.0.1",
    "langchain>=1.0.0",
    "chromadb>=1.2.0",
    "fastapi>=0.119.0",
    "pydantic>=2.12.3",
    "langchain-openai>=0.1.0",
    "requests>=2.32.0",
    "slowapi>=0.1.9",
    "boto3>=1.35.0",
    "duckduckgo-search>=6.3.0",
    "beautifulsoup4>=4.12.3",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "ruff",
    "anyio",
    "pytest",
    "langgraph-cli[inmem]",
]

[tool.setuptools]
packages = ["agent", "rag"]

[tool.setuptools.package-dir]
agent = "src/agent"
rag = "src/rag"
